Move EventEmitter to lib/events and rename to common.emitter
