+++ lib/misc.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const events = require('events');
-
 const { either } = require('./functional');
 
 const falseness = () => false;
@@ -111,20 +109,6 @@ const shuffle = (
   arr.sort(() => Math.random() - 0.5)
 );
 
-const eventEmitter = (
-  // EventEmitter with wildcard
-) => {
-  const ee = new events.EventEmitter();
-  const emit = ee.emit;
-  ee.emit = (...args) => {
-    const ar = args.slice(0);
-    ar.unshift('*');
-    emit.apply(ee, ar);
-    emit.apply(ee, args);
-  };
-  return ee;
-};
-
 const restLeft = (
   // Rest left, transfor function
   fn // (args, arg1..argN, callback) to (arg1..argN, ...args, callback)
@@ -151,7 +135,6 @@ module.exports = {
   sequence,
   random,
   shuffle,
-  eventEmitter,
   restLeft,
   requireEither
 };
